requirements.txt
annotated-types==0.7.0
anyio==4.4.0
assemblyai==0.33.0
azure-cognitiveservices-speech==1.40.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
jiter==0.5.0
moviepy==1.0.3
numpy==2.1.0
openai==1.42.0
pillow==10.4.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.0.1
